refactor(graph): update and add graph response DTOs for frontend compatibility.
