+++ src/main/java/com/luciano/music_graph/dto/ArtistSearchResult.java
@@ -3,6 +3,6 @@ package com.luciano.music_graph.dto;
 import java.util.List;
 
 public record ArtistSearchResult(
-        List<ArtistSearchData> artist
+        List<ArtistSearchData> nodes
 ) {
 }
