+++ src/main/java/com/luciano/music_graph/dto/graph/GraphResponse.java
@@ -1,0 +1,9 @@
+package com.luciano.music_graph.dto.graph;
+
+import java.util.List;
+
+public record GraphResponse(
+        List<Node> nodes,
+        List<Link> links
+) {
+}
