+++ src/main/java/com/luciano/music_graph/dto/graph/Link.java
@@ -1,0 +1,9 @@
+package com.luciano.music_graph.dto.graph;
+
+public record Link(
+        String source,
+        String target,
+        int weight,
+        boolean active
+) {
+}
